make specifying superuserPassword & secretKey more secure
currently, the superuser, superuserPassword & secretKey are directly pulled as values into the backend chart. This prevents us from using Kubernetes Secrets for these values and instead requires them to be hard coded in values.yaml. This change allows Kubernetes Secrets to be used similar to how externaldb envs are handled